+++ provider/pkg/await/statefulset_test.go
@@ -203,8 +203,7 @@ func Test_Apps_StatefulSet(t *testing.T) {
 		period := make(chan time.Time)
 		go test.do(statefulsets, pods, timeout)
 
-		err := awaiter.await(
-			&chanWatcher{results: statefulsets}, &chanWatcher{results: pods}, timeout, period)
+		err := awaiter.await(statefulsets, pods, timeout, period)
 		assert.Equal(t, test.expectedError, err, test.description)
 	}
 }
@@ -262,8 +261,7 @@ func Test_Apps_StatefulSet_MultipleUpdates(t *testing.T) {
 				awaiter.config.lastInputs = obj
 			})
 
-		err := awaiter.await(
-			&chanWatcher{results: statefulsets}, &chanWatcher{results: pods}, timeout, period)
+		err := awaiter.await(statefulsets, pods, timeout, period)
 		assert.Equal(t, test.firstExpectedError, err, test.description)
 
 		statefulsets = make(chan watch.Event)
@@ -273,8 +271,7 @@ func Test_Apps_StatefulSet_MultipleUpdates(t *testing.T) {
 		period = make(chan time.Time)
 		go test.secondUpdate(statefulsets, pods, timeout)
 
-		err = awaiter.await(
-			&chanWatcher{results: statefulsets}, &chanWatcher{results: pods}, timeout, period)
+		err = awaiter.await(statefulsets, pods, timeout, period)
 		assert.Equal(t, test.secondExpectedError, err, test.description)
 	}
 }