+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## HEAD (Unreleased)
+- Use shared informer for await logic for all resources (https://github.com/pulumi/pulumi-kubernetes/pull/1647)
 
 ## 3.5.0 (June 30, 2021)